+++ YEngine/D3D.h
@@ -13,21 +13,21 @@ private:
 	ID3D11Device* gDevice;
 	ID3D11DeviceContext* gDevCon;
 	HWND hwnd;
-	size_t wWidth;
-	size_t wHeight;
+	int wWidth;
+	int wHeight;
 public:
 	virtual void initializeWindow(HINSTANCE hInstance, int ShowWnd, int width, int height, bool windowed);
 	virtual void createSwapChain();
-	virtual void createVertexBuffer(ID3D11Buffer** gVertexBuffer, void* v, size_t& stride, size_t& offset, size_t numVertices);
-	virtual void createIndexBuffer(ID3D11Buffer** gIndexBuffer, void* data, size_t& numIndices);
+	virtual void createVertexBuffer(ID3D11Buffer** gVertexBuffer, void* v, size_t& stride, size_t& offset, int numVertices);
+	virtual void createIndexBuffer(ID3D11Buffer** gIndexBuffer, void* data, int& numIndices);
 	virtual void setVertexBuffer(ID3D11Buffer** gVertexBuffer, size_t& stride, size_t& offset);
-	virtual void setIndexBuffer(ID3D11Buffer* gIndexBuffer, size_t offset);
+	virtual void setIndexBuffer(ID3D11Buffer* gIndexBuffer, int offset);
 	virtual void createConstantBuffer(ID3D11Buffer ** gBuffer, int bufferSize);
 	virtual void mapBuffer(ID3D11Buffer** gBuffer, void* cbPtr, int structSize);
 	virtual void cleanup();
 
-	virtual size_t& GETwWidth();
-	virtual size_t& GETwHeight();
+	virtual int& GETwWidth();
+	virtual int& GETwHeight();
 	virtual HWND& GEThwnd();
 	virtual ID3D11Device*& GETgDevice();
 	virtual ID3D11DeviceContext*& GETgDevCon();
